message: Add a rudimentary test for message parsing.
This is a simple test that makes sure that the `parseMessage` function properly returns the details of an incoming message when used within `NewMessage`. I tried to test as many fields as possible with a barebones bot.

P.S.: I'm in the process of writing some plugins for my fork of the bot; if you'd like, I can possibly send a pull request with some examples if I come across writing a good one.